fix employee page, dialog, validate date of birth
