+++ frontend/src/api/employee/employeeAPI.js
@@ -7,11 +7,19 @@ export const employeeAPI = {
     },
     create: (value) => {
         const url = `/employee/create`;
-        return axiosClient.post(url, value);
+        return axiosClient.post(url, value, {
+            headers: {
+                "Content-Type": "application/json",
+            }
+        });
     },
     update: (value, id) => {
         const url = `/employee/${id}/update`;
-        return axiosClient.put(url, value);
+        return axiosClient.put(url, value, {
+            headers: {
+                "Content-Type": "application/json",
+            }
+        });
     },
     delete: (id) => {
         const url = `/employee/${id}/delete`;
