+++ frontend/src/api/order/orderAPI.js
@@ -3,6 +3,10 @@ import axiosClient from "../axiosClient";
 export const orderAPI = {
     update: (id) => {
         const url = `/order/${id}/update`;
-        return axiosClient.put(url);
+        return axiosClient.put(url,{
+            headers: {
+                "Content-Type": "application/json",
+            }
+        });
     },
 }
