+++ frontend/src/api/order/ordersAPI.js
@@ -3,7 +3,6 @@ import axiosClient from "../axiosClient";
 export const ordersAPI = {
     getAll: (status) => {
         const url = status === null? `/order`:`order?status=${status}`;
-        console.log("url", url);
         return axiosClient.get(url);
     },
 }
