+++ frontend/src/api/product/productAPI.js
@@ -7,11 +7,19 @@ export const productAPI = {
     },
     create: (value) => {
         const url = `/product/create`;
-        return axiosClient.post(url, value);
+        return axiosClient.post(url, value, {
+            headers: {
+                "Content-Type": "multipart/form-data",
+            }
+        });
     },
     update: (value, id) => {
         const url = `/product/${id}/update`;
-        return axiosClient.put(url, value);
+        return axiosClient.put(url, value, {
+            headers: {
+                "Content-Type": "multipart/form-data",
+            }
+        });
     },
     delete: (id) => {
         const url = `/product/${id}/delete`;
