+++ frontend/src/pages/admin/employee/employee_dialog/index.js
@@ -1,0 +1,4 @@
+export { FormDialog as FormDialog } from "./view_dialog";
+export { FormCreateDialog as FormCreateDialog } from "./create_dialog";
+export { FormEditDialog as FormEditDialog } from "./update_dialog";
+export { FormDeleteDialog as FormDeleteDialog } from "./delete_dialog";
